added one-hot.py to generate label array
one-hot encoding is how the cnn will keep track of each image after processing and
check whether it predicts correctly or not. This is the testing script to make sure
that it outputs in the correct format. Now it needs to be integrated with the list
of classes coming from the unique class labels in the data itself to make it flexible.