input csv output transposed np array of xyz
clean csv data is converted to transposed numpy array so that when processing
it is easier to understand the output on a timeline. Instead of imagining time
moving from top to bottom (top row to bottom row) we can imagine it moving from
left to right and the positive/negative values of xyz can be imagined on the y-axis.