input converter creates stacked array
data, once cleaned, can now be converted into the "image" arrays for
input into the CNN. all "images" must be of the same dimesions as in 25 rows of
data. Now I need to find a way to keep track of the timestamps of each piece and
associate it with corresponding behavior class label.